+++ src/flyrigloader/discovery/files.py
@@ -1,1489 +1,17 @@
-"""
-File discovery functionality.
-
-Advanced utilities for finding files based on patterns with dependency injection
-support for comprehensive testing and mocking capabilities.
-
-This module implements the discovery phase of the decoupled pipeline architecture,
-focusing exclusively on metadata extraction without data loading. Enhanced with
-Kedro integration capabilities, version-aware pattern matching, and catalog-specific
-metadata extraction for seamless pipeline integration.
-"""
-import os
-from collections.abc import Iterable as IterableABC, Mapping
-from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
-from pathlib import Path
-import re
-from datetime import datetime
-import fnmatch
-from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
-
-from semantic_version import Version
-from flyrigloader import logger
-from flyrigloader.discovery.enumeration import (
-    FileEnumerator,
-    _normalize_directory_argument,
-)
-from flyrigloader.discovery.patterns import PatternMatcher, match_files_to_patterns
-from flyrigloader.discovery.providers import (
-    DateTimeProvider,
-    FilesystemProvider,
-    StandardDateTimeProvider,
-    StandardFilesystemProvider,
-)
-from flyrigloader.discovery.stats import get_file_stats, attach_file_stats
-from flyrigloader.config.versioning import CURRENT_SCHEMA_VERSION
-from flyrigloader.config.yaml_config import (
-    get_dataset_info,
-    get_experiment_info,
-    get_extraction_patterns,
-    get_ignore_patterns,
-)
-
-
-def _convert_substring_to_glob(pattern: str) -> str:
-    """Convert a raw substring ignore pattern into a glob-compatible string."""
-
-    if not pattern:
-        return pattern
-
-    if any(char in pattern for char in ("*", "?")):
-        return pattern
-
-    if pattern == "._":
-        return "*._*"
-
-    if pattern.startswith('.'):
-        return f"{pattern}*"
-
-    return f"*{pattern}*"
-
-
-@dataclass
-class FileInfo:
-    """
-    Information about a discovered file without loading its content.
-    
-    Enhanced with Kedro-specific metadata fields for catalog integration and
-    version-aware discovery patterns. Supports seamless pipeline integration
-    with data lineage tracking and catalog-aware workflows.
-    """
-    
-    path: str
-    size: Optional[int] = None
-    mtime: Optional[float] = None
-    ctime: Optional[float] = None
-    creation_time: Optional[float] = None
-    extracted_metadata: Dict[str, Any] = field(default_factory=dict)
-    parsed_date: Optional[datetime] = None
-    
-    # Kedro-specific metadata fields for catalog integration
-    kedro_dataset_name: Optional[str] = None
-    kedro_version: Optional[str] = None
-    kedro_namespace: Optional[str] = None
-    kedro_tags: List[str] = field(default_factory=list)
-    catalog_metadata: Dict[str, Any] = field(default_factory=dict)
-    
-    # Version-aware discovery fields
-    schema_version: str = CURRENT_SCHEMA_VERSION
-    version_compatibility: Optional[Dict[str, bool]] = field(default_factory=dict)
-    
-    def get_kedro_dataset_path(self) -> Optional[str]:
-        """
-        Generate Kedro dataset path for catalog integration.
-        
-        Returns:
-            Optional[str]: Formatted dataset path for Kedro catalog, or None if not applicable
-        """
-        if self.kedro_dataset_name:
-            if self.kedro_namespace:
-                return f"{self.kedro_namespace}.{self.kedro_dataset_name}"
-            return self.kedro_dataset_name
-        return None
-    
-    def is_kedro_versioned(self) -> bool:
-        """
-        Check if this file follows Kedro versioning patterns.
-        
-        Returns:
-            bool: True if file has Kedro version metadata
-        """
-        return bool(self.kedro_version and self.kedro_version != "latest")
-    
-    def get_version_info(self) -> Dict[str, Any]:
-        """
-        Get comprehensive version information for this file.
-        
-        Returns:
-            Dict[str, Any]: Version information including schema and Kedro versions
-        """
-        return {
-            "schema_version": self.schema_version,
-            "kedro_version": self.kedro_version,
-            "version_compatibility": self.version_compatibility or {},
-            "is_versioned": self.is_kedro_versioned()
-        }
-
-
-@dataclass
-class FileStatistics:
-    """Statistics about the discovered files."""
-    
-    total_files: int
-    total_size: int
-    file_types: Dict[str, int] = field(default_factory=dict)
-    date_range: Optional[Tuple[datetime, datetime]] = None
-    discovery_time: Optional[float] = None
-
-
-@dataclass
-class FileManifest:
-    """
-    Container for discovered files metadata without loading actual data.
-    
-    This class represents the result of the discovery phase in the decoupled
-    pipeline architecture, containing only metadata about discovered files.
-    Enhanced with Kedro catalog integration capabilities and version-aware
-    discovery patterns for seamless pipeline workflows.
-    """
-    
-    files: List[FileInfo]
-    metadata: Dict[str, Any] = field(default_factory=dict)
-    statistics: Optional[FileStatistics] = None
-    
-    # Kedro integration fields
-    kedro_catalog_entries: Dict[str, Dict[str, Any]] = field(default_factory=dict)
-    kedro_pipeline_compatibility: bool = True
-    supported_kedro_versions: List[str] = field(default_factory=list)
-    
-    # Version management fields
-    manifest_version: str = CURRENT_SCHEMA_VERSION
-    discovery_metadata: Dict[str, Any] = field(default_factory=dict)
-    
-    def __post_init__(self):
-        """Initialize computed properties after instantiation."""
-        if self.statistics is None and self.files:
-            self._compute_statistics()
-    
-    def _compute_statistics(self) -> None:
-        """Compute basic statistics from the discovered files."""
-        if not self.files:
-            return
-        
-        total_files = len(self.files)
-        total_size = sum(f.size or 0 for f in self.files)
-        
-        # Count file types by extension
-        file_types = {}
-        dates = []
-        
-        for file_info in self.files:
-            # Extract extension
-            path = Path(file_info.path)
-            ext = path.suffix.lower()
-            file_types[ext] = file_types.get(ext, 0) + 1
-            
-            # Collect dates for range calculation
-            if file_info.parsed_date:
-                dates.append(file_info.parsed_date)
-        
-        # Compute date range
-        date_range = None
-        if dates:
-            dates.sort()
-            date_range = (dates[0], dates[-1])
-        
-        self.statistics = FileStatistics(
-            total_files=total_files,
-            total_size=total_size,
-            file_types=file_types,
-            date_range=date_range
-        )
-    
-    def get_files_by_type(self, extension: str) -> List[FileInfo]:
-        """Get files filtered by extension."""
-        return [f for f in self.files if f.path.lower().endswith(extension.lower())]
-    
-    def get_files_by_pattern(self, pattern: str) -> List[FileInfo]:
-        """Get files matching a specific pattern."""
-        return [f for f in self.files if fnmatch.fnmatch(Path(f.path).name, pattern)]
-    
-    def get_kedro_compatible_files(self) -> List[FileInfo]:
-        """Get files that are compatible with Kedro pipeline integration."""
-        return [f for f in self.files if f.kedro_dataset_name or self._is_kedro_pattern(f.path)]
-    
-    def get_versioned_files(self) -> List[FileInfo]:
-        """Get files that have Kedro versioning metadata."""
-        return [f for f in self.files if f.is_kedro_versioned()]
-    
-    def get_files_by_namespace(self, namespace: str) -> List[FileInfo]:
-        """Get files belonging to a specific Kedro namespace."""
-        return [f for f in self.files if f.kedro_namespace == namespace]
-    
-    def generate_kedro_catalog_entries(self) -> Dict[str, Dict[str, Any]]:
-        """
-        Generate Kedro catalog entries for discovered files.
-        
-        Returns:
-            Dict[str, Dict[str, Any]]: Kedro catalog entries with dataset configurations
-        """
-        catalog_entries = {}
-        
-        for file_info in self.files:
-            if file_info.kedro_dataset_name:
-                entry_name = file_info.get_kedro_dataset_path() or file_info.kedro_dataset_name
-                
-                catalog_entries[entry_name] = {
-                    'type': 'flyrigloader.kedro.datasets.FlyRigLoaderDataSet',
-                    'filepath': file_info.path,
-                    'metadata': file_info.catalog_metadata.copy(),
-                    'versioned': file_info.is_kedro_versioned(),
-                    'tags': file_info.kedro_tags.copy()
-                }
-                
-                if file_info.kedro_version:
-                    catalog_entries[entry_name]['version'] = file_info.kedro_version
-        
-        self.kedro_catalog_entries = catalog_entries
-        return catalog_entries
-    
-    def validate_kedro_compatibility(self, kedro_version: str = "0.18.0") -> Tuple[bool, List[str]]:
-        """
-        Validate compatibility with a specific Kedro version.
-        
-        Args:
-            kedro_version: Target Kedro version to validate against
-            
-        Returns:
-            Tuple[bool, List[str]]: (is_compatible, list_of_issues)
-        """
-        issues = []
-        
-        try:
-            target_version = Version(kedro_version)
-            min_supported = Version("0.18.0")
-            
-            if target_version < min_supported:
-                issues.append(f"Kedro version {kedro_version} is below minimum supported version {min_supported}")
-        except Exception as e:
-            issues.append(f"Invalid Kedro version format: {kedro_version}")
-        
-        # Check for Kedro-specific file patterns
-        kedro_files = self.get_kedro_compatible_files()
-        if not kedro_files and self.files:
-            issues.append("No Kedro-compatible files found in manifest")
-        
-        # Validate dataset names
-        for file_info in kedro_files:
-            if file_info.kedro_dataset_name and not self._is_valid_kedro_name(file_info.kedro_dataset_name):
-                issues.append(f"Invalid Kedro dataset name: {file_info.kedro_dataset_name}")
-        
-        self.kedro_pipeline_compatibility = len(issues) == 0
-        return self.kedro_pipeline_compatibility, issues
-    
-    def get_version_summary(self) -> Dict[str, Any]:
-        """
-        Get version information summary for the manifest.
-        
-        Returns:
-            Dict[str, Any]: Comprehensive version information
-        """
-        version_counts = {}
-        for file_info in self.files:
-            version = file_info.schema_version
-            version_counts[version] = version_counts.get(version, 0) + 1
-        
-        return {
-            "manifest_version": self.manifest_version,
-            "file_version_distribution": version_counts,
-            "kedro_compatible_count": len(self.get_kedro_compatible_files()),
-            "versioned_files_count": len(self.get_versioned_files()),
-            "discovery_metadata": self.discovery_metadata
-        }
-    
-    def _is_kedro_pattern(self, filepath: str) -> bool:
-        """Check if file path follows Kedro naming conventions."""
-        path = Path(filepath)
-        
-        # Check for Kedro versioning patterns
-        kedro_version_pattern = r'^\d{4}-\d{2}-\d{2}T\d{2}\.\d{2}\.\d{2}\.\d{3}Z$'
-        if re.search(kedro_version_pattern, path.parent.name):
-            return True
-        
-        # Check for Kedro dataset name patterns  
-        kedro_name_pattern = r'^[a-zA-Z][a-zA-Z0-9_]*(\.[a-zA-Z][a-zA-Z0-9_]*)*$'
-        if '.' in path.stem and re.match(kedro_name_pattern, path.stem):
-            return True
-        
-        return False
-    
-    def _is_valid_kedro_name(self, name: str) -> bool:
-        """Validate Kedro dataset name format."""
-        # Kedro dataset names must be valid Python identifiers with optional namespace dots
-        pattern = r'^[a-zA-Z][a-zA-Z0-9_]*(\.[a-zA-Z][a-zA-Z0-9_]*)*$'
-        return bool(re.match(pattern, name))
-
-
-class FileDiscoverer:
-    """
-    Class for discovering files and extracting metadata from file paths.
-    
-    This class implements the discovery phase of the decoupled pipeline architecture,
-    focusing exclusively on metadata extraction without data loading. It provides:
-    - Finding files based on patterns with configurable filesystem access
-    - Extracting metadata using configurable regex patterns
-    - Parsing dates from filenames with configurable datetime providers
-    - Collecting file statistics (size, modification time) with configurable providers
-    - Support for pluggable pattern matchers through registry-based extensibility
-    
-    The FileDiscoverer returns only metadata without loading actual file content,
-    enabling fast directory traversal and independent validation as part of the
-    refactored three-stage pipeline: discover → load → transform.
-    
-    Supports TST-REF-001 requirements for dependency injection patterns and 
-    registry-based pattern matcher extensibility.
-    """
-    
-    def __init__(
-        self,
-        extract_patterns: Optional[List[str]] = None,
-        parse_dates: bool = False,
-        include_stats: bool = False,
-        filesystem_provider: Optional[FilesystemProvider] = None,
-        pattern_matcher: Optional[PatternMatcher] = None,
-        datetime_provider: Optional[DateTimeProvider] = None,
-        stats_provider: Optional[Callable[[Union[str, Path]], Dict[str, Any]]] = None,
-        test_mode: bool = False,
-        # Kedro integration parameters
-        enable_kedro_metadata: bool = True,
-        kedro_namespace: Optional[str] = None,
-        kedro_tags: Optional[List[str]] = None,
-        # Version management parameters
-        schema_version: str = CURRENT_SCHEMA_VERSION,
-        version_aware_patterns: bool = True
-    ):
-        """
-        Initialize the FileDiscoverer with dependency injection support and Kedro integration.
-        
-        Args:
-            extract_patterns: Optional list of regex patterns to extract metadata from file paths
-            parse_dates: If True, attempt to parse dates from filenames
-            include_stats: If True, include file statistics (size, mtime, ctime, creation_time)
-            filesystem_provider: Optional filesystem provider for dependency injection (TST-REF-001)
-            pattern_matcher: Optional pre-configured pattern matcher for dependency injection (TST-REF-002)
-            datetime_provider: Optional datetime provider for dependency injection (F-016)
-            stats_provider: Optional statistics provider for dependency injection (F-016)
-            test_mode: If True, enables test-specific behavior for TST-REF-003 requirements
-            enable_kedro_metadata: If True, extract Kedro-specific metadata for catalog integration
-            kedro_namespace: Default Kedro namespace for discovered datasets
-            kedro_tags: Default tags to apply to discovered Kedro datasets
-            schema_version: Configuration schema version for version-aware discovery
-            version_aware_patterns: If True, apply version-aware pattern matching
-        """
-        logger.debug(f"Initializing FileDiscoverer with patterns={extract_patterns}, dates={parse_dates}, stats={include_stats}, kedro_enabled={enable_kedro_metadata}, version_aware={version_aware_patterns}, test_mode={test_mode}")
-        
-        self.extract_patterns = extract_patterns
-        self.parse_dates = parse_dates
-        self.include_stats = include_stats
-        self.test_mode = test_mode
-        
-        # Kedro integration settings
-        self.enable_kedro_metadata = enable_kedro_metadata
-        self.kedro_namespace = kedro_namespace
-        self.kedro_tags = kedro_tags or []
-        
-        # Version management settings
-        self.schema_version = schema_version
-        self.version_aware_patterns = version_aware_patterns
-        
-        # Dependency injection (TST-REF-001, F-016)
-        self.filesystem_provider = filesystem_provider or StandardFilesystemProvider()
-        self.datetime_provider = datetime_provider or StandardDateTimeProvider()
-        self.stats_provider = stats_provider or get_file_stats
-        self.file_enumerator = FileEnumerator(
-            filesystem_provider=self.filesystem_provider,
-            test_mode=self.test_mode,
-        )
-
-        logger.debug(f"Using filesystem provider: {type(self.filesystem_provider).__name__}")
-        logger.debug(f"Using datetime provider: {type(self.datetime_provider).__name__}")
-        logger.debug(
-            "Using file enumerator: %s",
-            type(self.file_enumerator).__name__,
-        )
-        logger.debug(f"Using stats provider: {self.stats_provider.__name__ if hasattr(self.stats_provider, '__name__') else type(self.stats_provider).__name__}")
-        
-        # Field names for pattern extraction (for backward compatibility)
-        # These are based on the test patterns and expectations
-        self.field_names = {
-            # For mouse pattern (standalone mouse files)
-            "mouse": ["animal", "date", "condition", "replicate"],
-            # For rat pattern (standalone rat files)
-            "rat": ["date", "animal", "condition", "replicate"],
-            # For experiment pattern (experiment files with animal types)
-            "exp": ["experiment_id", "animal", "condition"]
-        }
-        
-        # Create or use provided pattern matcher (TST-REF-002)
-        # Enhanced with registry-based extensibility for pluggable pattern matchers
-        if pattern_matcher:
-            logger.debug("Using provided pattern matcher (supports registry-based extensibility)")
-            self.pattern_matcher = pattern_matcher
-            self.named_extract_patterns = getattr(pattern_matcher, 'patterns', None)
-        elif extract_patterns:
-            logger.debug(f"Creating pattern matcher with {len(extract_patterns)} patterns")
-            self.named_extract_patterns = self._convert_to_named_patterns(extract_patterns)
-            self.pattern_matcher = PatternMatcher(self.named_extract_patterns)
-        else:
-            logger.debug("No pattern matcher configured - registry can be used for dynamic pattern registration")
-            self.pattern_matcher = None
-            self.named_extract_patterns = None
-        
-        # Common date formats to try when parsing dates
-        self.date_formats = [
-            # Standard format YYYYMMDD
-            "%Y%m%d",
-            # ISO format YYYY-MM-DD
-            "%Y-%m-%d",
-            # US format MM-DD-YYYY
-            "%m-%d-%Y",
-            # With timestamp YYYYMMDD_HHMMSS
-            "%Y%m%d_%H%M%S"
-        ]
-        
-        logger.debug(f"FileDiscoverer initialized successfully with {len(self.date_formats)} date formats")
-    
-    def _convert_to_named_patterns(self, patterns: List[str]) -> List[str]:
-        """
-        Convert traditional regex patterns with positional groups to patterns with named groups.
-        
-        This handles backward compatibility with existing patterns that use positional groups
-        based on known patterns for mice, rats, and experiments.
-        
-        Args:
-            patterns: List of regex patterns with positional groups
-            
-        Returns:
-            List of regex patterns with named groups
-        """
-        named_patterns = []
-        
-        for pattern in patterns:
-            # Handle the specific patterns from the test directly
-            if pattern == r".*/(mouse)_(\d{8})_(\w+)_(\d+)\.csv":
-                # Mouse pattern
-                named_pattern = r".*/(?P<animal>mouse)_(?P<date>\d{8})_(?P<condition>\w+)_(?P<replicate>\d+)\.csv"
-            elif pattern == r".*/(\d{8})_(rat)_(\w+)_(\d+)\.csv":
-                # Rat pattern
-                named_pattern = r".*/(?P<date>\d{8})_(?P<animal>rat)_(?P<condition>\w+)_(?P<replicate>\d+)\.csv"
-            elif pattern == r".*/(exp\d+)_(\w+)_(\w+)\.csv":
-                # Experiment pattern
-                named_pattern = r".*/(?P<experiment_id>exp\d+)_(?P<animal>\w+)_(?P<condition>\w+)\.csv"
-            else:
-                # For other patterns, keep as is (could be improved for more general cases)
-                named_pattern = pattern
-            
-            named_patterns.append(named_pattern)
-        
-        return named_patterns
-    
-    def find_files(
-        self,
-        directory: Union[str, List[str]],
-        pattern: str,
-        recursive: bool = False,
-        extensions: Optional[List[str]] = None,
-        ignore_patterns: Optional[List[str]] = None,
-        mandatory_substrings: Optional[List[str]] = None
-    ) -> List[str]:
-        """Find files matching the criteria using the configured enumerator."""
-        return self.file_enumerator.find_files(
-            directory=directory,
-            pattern=pattern,
-            recursive=recursive,
-            extensions=extensions,
-            ignore_patterns=ignore_patterns,
-            mandatory_substrings=mandatory_substrings,
-        )
-
-    def extract_metadata(self, files: List[str]) -> Dict[str, Dict[str, Any]]:
-        """
-        Extract metadata from file paths using configurable regex patterns.
-        
-        Enhanced with structured logging for Section 2.2.8 requirements.
-        
-        Args:
-            files: List of file paths to extract metadata from
-            
-        Returns:
-            Dictionary mapping file paths to metadata dictionaries
-        """
-        logger.debug(f"Extracting metadata from {len(files)} files")
-        
-        try:
-            # Use the configurable PatternMatcher for pattern extraction (TST-REF-002)
-            if self.pattern_matcher:
-                logger.debug("Using pattern matcher for metadata extraction")
-                result = {}
-                unmatched_count = 0
-                for file_path in files:
-                    metadata = self.pattern_matcher.match_all(file_path)
-                    if metadata is None:
-                        metadata = {}
-                        unmatched_count += 1
-                    metadata["path"] = file_path
-                    result[file_path] = metadata
-
-                if unmatched_count > 0:
-                    logger.debug(
-                        f"Added basic metadata for {unmatched_count} unmatched files"
-                    )
-            else:
-                logger.debug("No pattern matcher configured, creating basic metadata")
-                # Create basic metadata with just the path
-                result = {file_path: {"path": file_path} for file_path in files}
-            
-            # Process dates if requested using configurable datetime provider
-            if self.parse_dates:
-                logger.debug("Processing dates for extracted metadata")
-                for file_path, file_info in result.items():
-                    self._parse_date(file_path, file_info)
-            
-            # For backward compatibility with test_pattern_extraction, manually fix exp001_mouse_baseline.csv
-            # to not be counted as a mouse file
-            compatibility_fixes = 0
-            for path, info in result.items():
-                # If this is an experiment file with experiment_id and animal fields
-                if "experiment_id" in info and info.get("experiment_id", "").startswith("exp"):
-                    # Get the base filename for easier matching
-                    basename = Path(path).name
-                    
-                    # If this is specifically exp001_mouse_baseline.csv and animal is mouse
-                    if basename == "exp001_mouse_baseline.csv" and info.get("animal") == "mouse":
-                        info["animal"] = "exp_mouse"
-                        compatibility_fixes += 1
-                        logger.debug(f"Applied backward compatibility fix for {basename}")
-            
-            if compatibility_fixes > 0:
-                logger.debug(f"Applied {compatibility_fixes} backward compatibility fixes")
-            
-            logger.debug(f"Metadata extraction completed for {len(result)} files")
-            return result
-            
-        except Exception as e:
-            logger.error(f"Error in extract_metadata: {e}")
-            if self.test_mode:
-                logger.warning("Test mode enabled, returning basic metadata despite error")
-                return {file_path: {"path": file_path} for file_path in files}
-            raise
-    
-    def _parse_date(self, file_path: str, file_info: Dict[str, Any]) -> None:
-        """
-        Parse date information from a file path and update file_info using configurable datetime provider.
-        
-        Enhanced with dependency injection support for F-016 requirements.
-        
-        Args:
-            file_path: Path to extract date from
-            file_info: Dictionary to update with parsed date
-        """
-        logger.debug(f"Parsing date from file path: {file_path}")
-        
-        # First check if we already extracted a date field
-        date_str = file_info.get("date")
-        logger.debug(f"Date field from metadata: {date_str}")
-        
-        # If not, try to extract from filename
-        if not date_str:
-            # Look for date patterns in the filename
-            basename = Path(file_path).name
-            logger.debug(f"Extracting date from filename: {basename}")
-            
-            # Try to find a date in the filename using regex
-            date_patterns = [
-                r"(\d{8})",            # YYYYMMDD
-                r"(\d{4}-\d{2}-\d{2})", # YYYY-MM-DD
-                r"(\d{2}-\d{2}-\d{4})", # MM-DD-YYYY
-            ]
-            
-            for date_pattern in date_patterns:
-                if date_match := re.search(date_pattern, basename):
-                    date_str = date_match[1]
-                    logger.debug(f"Found date string '{date_str}' using pattern '{date_pattern}'")
-                    break
-        
-        # Try to parse the date if found using configurable datetime provider
-        if date_str:
-            logger.debug(f"Attempting to parse date string: {date_str}")
-            for fmt in self.date_formats:
-                try:
-                    # Use configurable datetime provider for F-016 requirements
-                    parsed_date = self.datetime_provider.strptime(date_str, fmt)
-                    file_info["parsed_date"] = parsed_date
-                    logger.debug(f"Successfully parsed date: {parsed_date} using format: {fmt}")
-                    break
-                except ValueError as e:
-                    logger.debug(f"Failed to parse with format '{fmt}': {e}")
-                    continue
-            else:
-                logger.warning(f"Could not parse date string '{date_str}' with any known format")
-        else:
-            logger.debug("No date string found to parse")
-    
-    def discover(
-        self,
-        directory: Union[str, List[str]],
-        pattern: str,
-        recursive: bool = False,
-        extensions: Optional[List[str]] = None,
-        ignore_patterns: Optional[List[str]] = None,
-        mandatory_substrings: Optional[List[str]] = None
-    ) -> Union[List[str], Dict[str, Dict[str, Any]]]:
-        """
-        Discover files and extract metadata based on the configured options.
-        
-        This method implements the discovery phase of the decoupled pipeline architecture,
-        focusing exclusively on metadata extraction without data loading. It enables fast
-        directory traversal and independent validation.
-        
-        Enhanced with structured logging for Section 2.2.8 requirements.
-        
-        Args:
-            directory: The directory or list of directories to search in
-            pattern: File pattern to match (glob format)
-            recursive: If True, search recursively through subdirectories
-            extensions: Optional list of file extensions to filter by (without the dot)
-            ignore_patterns: Optional list of glob patterns to ignore. Examples include "*temp*", "backup_*", 
-                             or any other pattern that matches files to be excluded.
-            mandatory_substrings: Optional list of substrings that must be present in files
-            
-        Returns:
-            If extract_patterns, parse_dates, or include_stats is configured: Dictionary mapping 
-            file paths to extracted metadata (NO data loading occurs).
-            Otherwise: List of matched file paths.
-        """
-        logger.debug(f"Starting file discovery process with pattern: {pattern}")
-        logger.debug(f"Discovery options - patterns: {bool(self.extract_patterns)}, dates: {self.parse_dates}, stats: {self.include_stats}")
-        
-        try:
-            # Find files matching the criteria
-            found_files = self.find_files(
-                directory,
-                pattern,
-                recursive=recursive,
-                extensions=extensions,
-                ignore_patterns=ignore_patterns,
-                mandatory_substrings=mandatory_substrings
-            )
-            
-            # Check if we need to return files with metadata
-            if not (self.extract_patterns or self.parse_dates or self.include_stats or self.enable_kedro_metadata):
-                logger.debug(f"Returning simple file list with {len(found_files)} files")
-                return found_files
-            
-            logger.debug("Extracting metadata and additional information")
-            # Extract metadata from paths and return the result
-            result = self._extract_metadata_from_paths(found_files)
-
-            # Add file statistics if requested using configurable provider
-            if self.include_stats:
-                logger.debug("Adding file statistics to results")
-                try:
-                    # Use configurable stats provider for F-016 requirements
-                    if self.stats_provider == get_file_stats:
-                        # Use the standard attach_file_stats function
-                        result = attach_file_stats(result)
-                    else:
-                        # Use custom stats provider
-                        for file_path, metadata in result.items():
-                            try:
-                                stats = self.stats_provider(file_path)
-                                metadata.update(stats)
-                            except Exception as e:
-                                logger.error(f"Error getting stats for {file_path}: {e}")
-                                if not self.test_mode:
-                                    raise
-                    logger.debug("File statistics added successfully")
-                except Exception as e:
-                    logger.error(f"Error adding file statistics: {e}")
-                    if not self.test_mode:
-                        raise
-            
-            logger.debug(f"Discovery completed successfully with {len(result)} files")
-            return result
-            
-        except Exception as e:
-            logger.error(f"Critical error in discover method: {e}")
-            if self.test_mode:
-                logger.warning("Test mode enabled, returning empty result despite error")
-                return [] if not (self.extract_patterns or self.parse_dates or self.include_stats) else {}
-            raise
-    
-    def _extract_metadata_from_paths(self, files: List[str]) -> Dict[str, Dict[str, Any]]:
-        """
-        Extract metadata from file paths with version-aware and Kedro integration.
-        
-        Args:
-            files: List of file paths to extract metadata from
-            
-        Returns:
-            Dictionary mapping file paths to metadata dictionaries
-        """
-        # Apply version-aware filtering if enabled
-        if self.version_aware_patterns:
-            files = self.apply_version_aware_patterns(files)
-        
-        # Extract base metadata
-        base_metadata = self.extract_metadata(files)
-        
-        # Enhance with Kedro metadata if enabled
-        if self.enable_kedro_metadata:
-            for file_path, metadata in base_metadata.items():
-                kedro_metadata = self.extract_kedro_metadata(file_path)
-                metadata.update(kedro_metadata)
-        
-        return base_metadata
-    
-    def register_pattern_matcher(self, pattern_matcher: PatternMatcher) -> None:
-        """
-        Register a new pattern matcher for enhanced pluggable pattern matching.
-        
-        This method supports registry-based extensibility for pluggable pattern matchers
-        as mentioned in Section 0.2.1 of the refactoring requirements.
-        
-        Args:
-            pattern_matcher: The pattern matcher instance to register
-        """
-        logger.debug(f"Registering new pattern matcher: {type(pattern_matcher).__name__}")
-        self.pattern_matcher = pattern_matcher
-        self.named_extract_patterns = getattr(pattern_matcher, 'patterns', None)
-        logger.debug(f"Pattern matcher registered successfully: {type(pattern_matcher).__name__}")
-    
-    def get_supported_patterns(self) -> List[str]:
-        """
-        Get list of supported patterns from the current pattern matcher.
-        
-        Returns:
-            List of supported pattern names, or empty list if no matcher is configured
-        """
-        if self.pattern_matcher and hasattr(self.pattern_matcher, 'patterns'):
-            return list(self.pattern_matcher.patterns.keys()) if isinstance(self.pattern_matcher.patterns, dict) else []
-        return []
-    
-    def extract_kedro_metadata(self, file_path: str) -> Dict[str, Any]:
-        """
-        Extract Kedro-specific metadata from file path for catalog integration.
-        
-        Args:
-            file_path: Path to extract Kedro metadata from
-            
-        Returns:
-            Dict[str, Any]: Kedro-specific metadata including dataset name, version, namespace
-        """
-        if not self.enable_kedro_metadata:
-            return {}
-        
-        logger.debug(f"Extracting Kedro metadata from: {file_path}")
-        
-        path = Path(file_path)
-        kedro_metadata = {}
-        
-        # Extract Kedro dataset name from filename or directory structure
-        dataset_name = self._extract_kedro_dataset_name(path)
-        if dataset_name:
-            kedro_metadata['kedro_dataset_name'] = dataset_name
-            logger.debug(f"Detected Kedro dataset name: {dataset_name}")
-        
-        # Extract Kedro version information
-        version_info = self._extract_kedro_version_info(path)
-        if version_info:
-            kedro_metadata.update(version_info)
-            logger.debug(f"Detected Kedro version info: {version_info}")
-        
-        # Apply default namespace and tags
-        if self.kedro_namespace:
-            kedro_metadata['kedro_namespace'] = self.kedro_namespace
-        
-        if self.kedro_tags:
-            kedro_metadata['kedro_tags'] = self.kedro_tags.copy()
-        
-        # Extract catalog-specific attributes
-        catalog_attrs = self._extract_catalog_attributes(path)
-        if catalog_attrs:
-            kedro_metadata['catalog_metadata'] = catalog_attrs
-        
-        return kedro_metadata
-    
-    def _extract_kedro_dataset_name(self, path: Path) -> Optional[str]:
-        """Extract Kedro dataset name from file path."""
-        # Check for explicit dataset name in parent directory
-        if path.parent.name.startswith('dataset_'):
-            return path.parent.name[8:]  # Remove 'dataset_' prefix
-        
-        # Check for namespaced dataset names (namespace.dataset format)
-        stem = path.stem
-        if '.' in stem and not stem.startswith('.'):
-            # Validate as potential Kedro dataset name
-            if re.match(r'^[a-zA-Z][a-zA-Z0-9_]*(\.[a-zA-Z][a-zA-Z0-9_]*)*$', stem):
-                return stem
-        
-        # Generate dataset name from filename (experiment-based)
-        if 'experiment' in stem or 'exp' in stem:
-            # Clean up common patterns for experiment names
-            clean_name = re.sub(r'[^a-zA-Z0-9_.]', '_', stem)
-            clean_name = re.sub(r'_+', '_', clean_name).strip('_')
-            if clean_name and clean_name[0].isalpha():
-                return clean_name
-        
-        return None
-    
-    def _extract_kedro_version_info(self, path: Path) -> Dict[str, Any]:
-        """Extract Kedro version information from file path."""
-        version_info = {}
-        
-        # Check for Kedro timestamp versioning pattern in parent directories
-        for parent in path.parents:
-            # Kedro timestamp pattern: YYYY-MM-DDTHH.MM.SS.fffZ
-            timestamp_pattern = r'^(\d{4}-\d{2}-\d{2}T\d{2}\.\d{2}\.\d{2}\.\d{3}Z)$'
-            match = re.match(timestamp_pattern, parent.name)
-            if match:
-                version_info['kedro_version'] = match.group(1)
-                break
-        
-        # Check for semantic version pattern in path
-        version_pattern = r'v?(\d+\.\d+\.\d+(?:-[a-zA-Z0-9.]+)?)'
-        version_match = re.search(version_pattern, str(path))
-        if version_match and 'kedro_version' not in version_info:
-            version_info['kedro_version'] = version_match.group(1)
-        
-        return version_info
-    
-    def _extract_catalog_attributes(self, path: Path) -> Dict[str, Any]:
-        """Extract catalog-specific attributes from file path."""
-        catalog_attrs = {}
-        
-        # Detect file format and size for catalog optimization
-        catalog_attrs['file_format'] = path.suffix.lower()
-        
-        # Detect experimental metadata that might be useful for catalog
-        path_str = str(path).lower()
-        
-        if 'baseline' in path_str:
-            catalog_attrs['experiment_type'] = 'baseline'
-        elif 'treatment' in path_str:
-            catalog_attrs['experiment_type'] = 'treatment'
-        elif 'control' in path_str:
-            catalog_attrs['experiment_type'] = 'control'
-        
-        # Extract animal type if present
-        for animal_type in ['mouse', 'rat', 'fly']:
-            if animal_type in path_str:
-                catalog_attrs['animal_type'] = animal_type
-                break
-        
-        return catalog_attrs
-    
-    def apply_version_aware_patterns(self, files: List[str], target_version: Optional[str] = None) -> List[str]:
-        """
-        Apply version-aware pattern matching based on configuration schema versions.
-        
-        Args:
-            files: List of file paths to filter
-            target_version: Target schema version for compatibility filtering
-            
-        Returns:
-            List[str]: Filtered files compatible with the target version
-        """
-        if not self.version_aware_patterns:
-            return files
-        
-        if target_version is None:
-            target_version = self.schema_version
-        
-        logger.debug(f"Applying version-aware patterns for version {target_version}")
-        
-        try:
-            target_ver = Version(target_version)
-            current_ver = Version(CURRENT_SCHEMA_VERSION)
-            
-            # If target version is current or newer, no filtering needed
-            if target_ver >= current_ver:
-                return files
-            
-            # Apply version-specific filtering
-            return self._filter_files_by_version_compatibility(files, target_ver)
-            
-        except Exception as e:
-            logger.warning(f"Version-aware pattern matching failed: {e}")
-            return files
-    
-    def _filter_files_by_version_compatibility(self, files: List[str], target_version: Version) -> List[str]:
-        """Filter files based on version compatibility."""
-        compatible_files = []
-        
-        for file_path in files:
-            # Check if file follows patterns compatible with target version
-            if self._is_version_compatible(file_path, target_version):
-                compatible_files.append(file_path)
-        
-        logger.debug(f"Version filtering: {len(files)} -> {len(compatible_files)} files")
-        return compatible_files
-    
-    def _is_version_compatible(self, file_path: str, target_version: Version) -> bool:
-        """Check if a file is compatible with the target version."""
-        # Legacy version patterns (0.x.x) are more restrictive
-        if target_version.major == 0:
-            # For 0.x versions, avoid files with modern Kedro patterns
-            if 'dataset_' in file_path or re.search(r'\d{4}-\d{2}-\d{2}T\d{2}\.\d{2}\.\d{2}', file_path):
-                return False
-        
-        # Modern versions (1.x.x) support all patterns
-        return True
-    
-    def create_version_aware_file_info(self, file_path: str, metadata: Dict[str, Any]) -> FileInfo:
-        """
-        Create FileInfo with version-aware metadata and Kedro integration.
-        
-        Args:
-            file_path: Path to the file
-            metadata: Extracted metadata dictionary
-            
-        Returns:
-            FileInfo: Enhanced FileInfo object with version and Kedro metadata
-        """
-        # Extract Kedro metadata if enabled
-        kedro_metadata = self.extract_kedro_metadata(file_path) if self.enable_kedro_metadata else {}
-
-        # Normalize metadata payload to avoid mutating the caller's dictionary
-        normalized_metadata = dict(metadata) if metadata else {}
-
-        def _pop_stat(*keys: str) -> Any:
-            value: Any = None
-            found = False
-            for key in keys:
-                if key in normalized_metadata:
-                    if not found:
-                        value = normalized_metadata[key]
-                        found = True
-                    normalized_metadata.pop(key)
-            return value
-
-        size = _pop_stat('size', 'size_bytes')
-        mtime = _pop_stat('mtime', 'modified_time')
-        ctime = _pop_stat('ctime')
-        creation_time = _pop_stat('creation_time', 'created_time')
-        parsed_date = normalized_metadata.pop('parsed_date', None)
-
-        # Create base FileInfo populated with extracted statistics
-        file_info = FileInfo(
-            path=file_path,
-            size=size,
-            mtime=mtime,
-            ctime=ctime,
-            creation_time=creation_time,
-            extracted_metadata=normalized_metadata,
-            parsed_date=parsed_date,
-            schema_version=self.schema_version
-        )
-        
-        # Add Kedro-specific fields
-        if kedro_metadata:
-            file_info.kedro_dataset_name = kedro_metadata.get('kedro_dataset_name')
-            file_info.kedro_version = kedro_metadata.get('kedro_version')
-            file_info.kedro_namespace = kedro_metadata.get('kedro_namespace')
-            file_info.kedro_tags = kedro_metadata.get('kedro_tags', [])
-            file_info.catalog_metadata = kedro_metadata.get('catalog_metadata', {})
-        
-        # Add version compatibility information
-        file_info.version_compatibility = self._get_version_compatibility_info(file_path)
-        
-        return file_info
-    
-    def _get_version_compatibility_info(self, file_path: str) -> Dict[str, bool]:
-        """Get version compatibility information for a file."""
-        compatibility = {}
-        
-        try:
-            current_ver = Version(CURRENT_SCHEMA_VERSION)
-            
-            # Check compatibility with major versions
-            for major_version in [0, 1]:
-                test_version = Version(f"{major_version}.0.0")
-                compatibility[f"v{major_version}.x"] = self._is_version_compatible(file_path, test_version)
-            
-            # Check Kedro compatibility
-            has_kedro_patterns = self._has_kedro_patterns(file_path)
-            compatibility["kedro_compatible"] = has_kedro_patterns or self.enable_kedro_metadata
-            
-        except Exception as e:
-            logger.debug(f"Error determining version compatibility for {file_path}: {e}")
-        
-        return compatibility
-    
-    def _has_kedro_patterns(self, file_path: str) -> bool:
-        """Check if file path contains Kedro-specific patterns."""
-        path = Path(file_path)
-        
-        # Check for Kedro versioning patterns
-        if re.search(r'\d{4}-\d{2}-\d{2}T\d{2}\.\d{2}\.\d{2}', str(path)):
-            return True
-        
-        # Check for dataset naming patterns
-        if 'dataset_' in str(path):
-            return True
-        
-        # Check for namespace patterns in filename
-        if '.' in path.stem and re.match(r'^[a-zA-Z][a-zA-Z0-9_]*\.[a-zA-Z][a-zA-Z0-9_]*', path.stem):
-            return True
-        
-        return False
-
-
-def discover_experiment_manifest(
-    config: Union[Dict[str, Any], Any],
-    experiment_name: str,
-    patterns: Optional[List[str]] = None,
-    parse_dates: bool = True,
-    include_stats: bool = True,
-    filesystem_provider: Optional[FilesystemProvider] = None,
-    pattern_matcher: Optional[PatternMatcher] = None,
-    datetime_provider: Optional[DateTimeProvider] = None,
-    test_mode: bool = False,
-    # Kedro integration parameters
-    enable_kedro_metadata: bool = True,
-    kedro_namespace: Optional[str] = None,
-    kedro_tags: Optional[List[str]] = None,
-    # Version management parameters
-    schema_version: Optional[str] = None,
-    version_aware_patterns: bool = True
-) -> FileManifest:
-    """
-    Discover experiment files and return metadata-only manifest with Kedro integration.
-    
-    This function implements the discovery phase of the decoupled pipeline architecture,
-    focusing exclusively on metadata extraction without data loading. Enhanced with
-    Kedro catalog integration, version-aware discovery patterns, and catalog-specific
-    metadata extraction for seamless pipeline workflows.
-    
-    Args:
-        config: Configuration object or dictionary containing discovery settings
-        experiment_name: Name of the experiment to discover files for
-        patterns: Optional list of regex patterns to extract metadata from filenames
-        parse_dates: If True, attempt to parse dates from filenames
-        include_stats: If True, include file statistics in the manifest
-        filesystem_provider: Optional filesystem provider for testing scenarios
-        pattern_matcher: Optional pre-configured pattern matcher
-        datetime_provider: Optional datetime provider for testing scenarios
-        test_mode: If True, enables test-specific behavior
-        enable_kedro_metadata: If True, extract Kedro-specific metadata for catalog integration
-        kedro_namespace: Default Kedro namespace for discovered datasets
-        kedro_tags: Default tags to apply to discovered Kedro datasets
-        schema_version: Configuration schema version for version-aware discovery
-        version_aware_patterns: If True, apply version-aware pattern matching
-    
-    Returns:
-        FileManifest containing discovered files metadata with Kedro integration support
-    """
-    logger.debug(f"Discovering experiment manifest for '{experiment_name}'")
-
-    try:
-        # Normalize configuration input
-        if hasattr(config, "model_dump") and not isinstance(config, Mapping):
-            config_mapping: Mapping[str, Any] = config.model_dump()  # type: ignore[assignment]
-        else:
-            config_mapping = config  # type: ignore[assignment]
-
-        if not isinstance(config_mapping, Mapping):
-            raise TypeError(
-                "Configuration for discover_experiment_manifest must be mapping-like or support model_dump()."
-            )
-
-        if "project" not in config_mapping:
-            raise KeyError("Configuration missing 'project' section required for discovery")
-
-        project_config = config_mapping["project"]
-        if not isinstance(project_config, Mapping):
-            raise TypeError("Project configuration must be mapping-like for discovery operations")
-
-        directories_cfg = project_config.get("directories", {})
-        if not isinstance(directories_cfg, Mapping):
-            raise TypeError("Project directories configuration must be mapping-like")
-
-        base_directory = directories_cfg.get("major_data_directory")
-        if not base_directory:
-            raise ValueError("Project configuration must define 'major_data_directory' for discovery")
-
-        base_directory_path = Path(str(base_directory))
-
-        # Gather experiment and dataset information from config helpers
-        experiment_info = get_experiment_info(config_mapping, experiment_name)
-        dataset_names = experiment_info.get("datasets", [])
-        if not dataset_names:
-            raise ValueError(
-                f"Experiment '{experiment_name}' does not reference any datasets; unable to perform discovery."
-            )
-
-        extraction_patterns = patterns or get_extraction_patterns(config_mapping, experiment_name)
-
-        ignore_patterns = get_ignore_patterns(config_mapping, experiment_name)
-
-        project_extensions = project_config.get("file_extensions")
-        if project_extensions is None:
-            extensions: Optional[List[str]] = None
-        elif isinstance(project_extensions, IterableABC) and not isinstance(project_extensions, (str, bytes)):
-            extensions = [str(ext) for ext in project_extensions]
-        else:
-            raise TypeError("project.file_extensions must be a list of extensions when provided")
-
-        recursive = True
-
-        dataset_targets: List[Dict[str, Any]] = []
-        dataset_ignore_patterns: List[str] = []
-
-        for dataset_name in dataset_names:
-            dataset_info = get_dataset_info(config_mapping, dataset_name)
-
-            raw_patterns = dataset_info.get("patterns") or ["*"]
-            if isinstance(raw_patterns, IterableABC) and not isinstance(raw_patterns, (str, bytes)):
-                dataset_patterns = [str(pattern) for pattern in raw_patterns]
-            else:
-                raise TypeError(
-                    f"Dataset '{dataset_name}' patterns must be provided as an iterable of strings"
-                )
-
-            dataset_filters = dataset_info.get("filters", {})
-            if dataset_filters and not isinstance(dataset_filters, Mapping):
-                raise TypeError(f"Dataset '{dataset_name}' filters must be mapping-like if provided")
-
-            if isinstance(dataset_filters, Mapping):
-                ignore_from_dataset = dataset_filters.get("ignore_substrings", [])
-                if ignore_from_dataset:
-                    if not isinstance(ignore_from_dataset, IterableABC) or isinstance(
-                        ignore_from_dataset, (str, bytes)
-                    ):
-                        raise TypeError(
-                            f"Dataset '{dataset_name}' ignore_substrings must be a list of strings"
-                        )
-                    dataset_ignore_patterns.extend(
-                        _convert_substring_to_glob(str(pattern)) for pattern in ignore_from_dataset
-                    )
-
-            dataset_base_dir = base_directory_path / dataset_name
-            directories: List[str] = []
-
-            dates_vials = dataset_info.get("dates_vials")
-            if isinstance(dates_vials, Mapping) and dates_vials:
-                for date_key in dates_vials.keys():
-                    directories.append(str(dataset_base_dir / str(date_key)))
-            else:
-                directories.append(str(dataset_base_dir))
-
-            dataset_targets.append(
-                {
-                    "dataset": dataset_name,
-                    "directories": directories,
-                    "patterns": dataset_patterns,
-                }
-            )
-
-        combined_ignore_patterns = list(dict.fromkeys(ignore_patterns + dataset_ignore_patterns))
-
-        # Determine schema version from config or use provided/default
-        config_schema_version = schema_version
-        if config_schema_version is None:
-            config_schema_version = getattr(config, 'schema_version', CURRENT_SCHEMA_VERSION)
-        
-        logger.debug(f"Using schema version: {config_schema_version}")
-        
-        # Create FileDiscoverer with enhanced Kedro and version support
-        discoverer = FileDiscoverer(
-            extract_patterns=extraction_patterns,
-            parse_dates=parse_dates,
-            include_stats=include_stats,
-            filesystem_provider=filesystem_provider,
-            pattern_matcher=pattern_matcher,
-            datetime_provider=datetime_provider,
-            test_mode=test_mode,
-            enable_kedro_metadata=enable_kedro_metadata,
-            kedro_namespace=kedro_namespace or experiment_name,
-            kedro_tags=kedro_tags,
-            schema_version=config_schema_version,
-            version_aware_patterns=version_aware_patterns
-        )
-        
-        # Discover files for each pattern
-        all_files = []
-        seen_files: Set[str] = set()
-        experiment_metadata = {
-            'experiment_name': experiment_name,
-            'base_directory': str(base_directory_path),
-            'search_targets': dataset_targets,
-            'discovery_settings': {
-                'recursive': recursive,
-                'parse_dates': parse_dates,
-                'include_stats': include_stats,
-                'extensions': extensions,
-                'ignore_patterns': combined_ignore_patterns,
-                'extraction_patterns': extraction_patterns or [],
-            },
-            'ignore_patterns': combined_ignore_patterns,
-            'datasets': dataset_names,
-        }
-
-        for target in dataset_targets:
-            dataset_name = target['dataset']
-            for directory in target['directories']:
-                for pattern in target['patterns']:
-                    logger.debug(
-                        f"Discovering files for dataset '{dataset_name}' in '{directory}' with pattern '{pattern}'"
-                    )
-
-                    discovered = discoverer.discover(
-                        directory=str(directory),
-                        pattern=pattern,
-                        recursive=recursive,
-                        extensions=extensions,
-                        ignore_patterns=combined_ignore_patterns
-                    )
-
-                    if isinstance(discovered, dict):
-                        for file_path, metadata in discovered.items():
-                            if file_path in seen_files:
-                                continue
-                            seen_files.add(file_path)
-
-                            file_info = discoverer.create_version_aware_file_info(file_path, metadata)
-
-                            all_files.append(file_info)
-                    else:
-                        for file_path in discovered:
-                            if file_path in seen_files:
-                                continue
-                            seen_files.add(file_path)
-
-                            file_info = discoverer.create_version_aware_file_info(file_path, {})
-                            all_files.append(file_info)
-
-        # Create enhanced FileManifest with Kedro integration
-        manifest = FileManifest(
-            files=all_files,
-            metadata=experiment_metadata,
-            manifest_version=config_schema_version
-        )
-
-        # Add discovery metadata for version tracking
-        manifest.discovery_metadata = {
-            'discovery_timestamp': datetime.now().isoformat(),
-            'schema_version': config_schema_version,
-            'kedro_enabled': enable_kedro_metadata,
-            'version_aware_patterns': version_aware_patterns,
-            'experiment_name': experiment_name,
-            'datasets': dataset_names,
-        }
-        
-        # Generate Kedro catalog entries if enabled
-        if enable_kedro_metadata:
-            try:
-                catalog_entries = manifest.generate_kedro_catalog_entries()
-                logger.debug(f"Generated {len(catalog_entries)} Kedro catalog entries")
-                
-                # Validate Kedro compatibility
-                is_compatible, issues = manifest.validate_kedro_compatibility()
-                if not is_compatible:
-                    logger.warning(f"Kedro compatibility issues found: {issues}")
-                
-            except Exception as e:
-                logger.warning(f"Error generating Kedro catalog entries: {e}")
-        
-        logger.debug(f"Created enhanced manifest with {len(all_files)} files for experiment '{experiment_name}'")
-        
-        # Log version and Kedro statistics
-        version_summary = manifest.get_version_summary()
-        logger.debug(f"Manifest version summary: {version_summary}")
-        
-        return manifest
-        
-    except Exception as e:
-        logger.error(f"Error discovering experiment manifest for '{experiment_name}': {e}")
-        if test_mode:
-            logger.warning("Test mode enabled, returning empty manifest despite error")
-            return FileManifest(files=[], metadata={'experiment_name': experiment_name, 'error': str(e)})
-        raise
-
-
-def discover_files(
-    directory: Union[str, List[str]], 
-    pattern: str,
-    recursive: bool = False,
-    extensions: Optional[List[str]] = None,
-    ignore_patterns: Optional[List[str]] = None,
-    mandatory_substrings: Optional[List[str]] = None,
-    extract_patterns: Optional[List[str]] = None,
-    parse_dates: bool = False,
-    include_stats: bool = False,
-    # Test-specific entry points for TST-REF-003 requirements
-    filesystem_provider: Optional[FilesystemProvider] = None,
-    pattern_matcher: Optional[PatternMatcher] = None,
-    datetime_provider: Optional[DateTimeProvider] = None,
-    stats_provider: Optional[Callable[[Union[str, Path]], Dict[str, Any]]] = None,
-    test_mode: bool = False,
-    # Kedro integration parameters
-    enable_kedro_metadata: bool = False,
-    kedro_namespace: Optional[str] = None,
-    kedro_tags: Optional[List[str]] = None,
-    # Version management parameters
-    schema_version: str = CURRENT_SCHEMA_VERSION,
-    version_aware_patterns: bool = False
-) -> Union[List[str], Dict[str, Dict[str, Any]]]:
-    """
-    Discover files matching the given pattern and criteria with Kedro integration support.
-    
-    This function implements the discovery phase of the decoupled pipeline architecture,
-    focusing exclusively on metadata extraction without data loading. Enhanced with
-    Kedro-specific metadata extraction, version-aware pattern matching, and catalog
-    integration capabilities for seamless pipeline workflows.
-    
-    For structured experiment discovery with full Kedro integration, consider using
-    discover_experiment_manifest() which returns comprehensive FileManifest objects.
-    
-    Args:
-        directory: Directory or list of directories to search
-        pattern: Glob pattern to match
-        recursive: Whether to search recursively
-        extensions: Optional list of file extensions to filter by
-        ignore_patterns: Optional list of glob patterns to ignore. Examples include "*temp*", "backup_*", 
-                         or any other pattern that matches files to be excluded.
-        mandatory_substrings: Optional list of substrings that must be present
-        extract_patterns: Optional list of regex patterns to extract metadata
-        parse_dates: If True, attempt to parse dates from filenames
-        include_stats: If True, include file statistics (size, mtime, ctime, creation_time)
-        
-        # Test-specific parameters for dependency injection (TST-REF-003, F-016)
-        filesystem_provider: Optional filesystem provider for testing scenarios
-        pattern_matcher: Optional pre-configured pattern matcher for testing
-        datetime_provider: Optional datetime provider for testing scenarios
-        stats_provider: Optional statistics provider for testing scenarios
-        test_mode: If True, enables test-specific behavior and error handling
-        
-        # Kedro integration parameters
-        enable_kedro_metadata: If True, extract Kedro-specific metadata for catalog integration
-        kedro_namespace: Default Kedro namespace for discovered datasets
-        kedro_tags: Default tags to apply to discovered Kedro datasets
-        
-        # Version management parameters
-        schema_version: Configuration schema version for version-aware discovery
-        version_aware_patterns: If True, apply version-aware pattern matching
-    
-    Returns:
-        If extract_patterns, parse_dates, include_stats, or enable_kedro_metadata is used: 
-        Dictionary mapping file paths to extracted metadata (NO data loading occurs).
-        Otherwise: List of matched file paths.
-    """
-    logger.debug(f"discover_files called with pattern='{pattern}', test_mode={test_mode}")
-    
-    try:
-        normalized_directory = _normalize_directory_argument(directory)
-        logger.debug(f"Forwarding normalized directories: {normalized_directory}")
-
-        discoverer = FileDiscoverer(
-            extract_patterns=extract_patterns,
-            parse_dates=parse_dates,
-            include_stats=include_stats,
-            filesystem_provider=filesystem_provider,
-            pattern_matcher=pattern_matcher,
-            datetime_provider=datetime_provider,
-            stats_provider=stats_provider,
-            test_mode=test_mode,
-            enable_kedro_metadata=enable_kedro_metadata,
-            kedro_namespace=kedro_namespace,
-            kedro_tags=kedro_tags,
-            schema_version=schema_version,
-            version_aware_patterns=version_aware_patterns
-        )
-        
-        result = discoverer.discover(
-            normalized_directory,
-            pattern,
-            recursive=recursive,
-            extensions=extensions,
-            ignore_patterns=ignore_patterns,
-            mandatory_substrings=mandatory_substrings
-        )
-        
-        logger.debug(f"discover_files completed successfully, returned {len(result) if isinstance(result, (list, dict)) else 'unknown'} items")
-        return result
-        
-    except Exception as e:
-        logger.error(f"Error in discover_files: {e}")
-        if test_mode:
-            logger.warning("Test mode enabled, continuing despite error")
-            return [] if not any([extract_patterns, parse_dates, include_stats, enable_kedro_metadata]) else {}
-        raise
-
-
-def get_latest_file(
-    files: List[str],
-    # Test-specific entry points for TST-REF-003 requirements
-    filesystem_provider: Optional[FilesystemProvider] = None,
-    test_mode: bool = False
-) -> Optional[str]:
-    """
-    Return the most recently modified file from the list with test hook support.
-    
-    This utility function supports the discovery phase of the decoupled pipeline
-    architecture by providing file selection based on modification time metadata.
-    
-    Enhanced with test-specific entry points for TST-REF-003 requirements enabling
-    controlled behavior during test execution.
-    
-    Args:
-        files: List of file paths to evaluate
-        filesystem_provider: Optional filesystem provider for testing scenarios (F-016)
-        test_mode: If True, enables test-specific behavior and error handling
-        
-    Returns:
-        Path to the most recently modified file, or None if no files provided
-    """
-    logger.debug(f"get_latest_file called with {len(files) if files else 0} files, test_mode={test_mode}")
-    
-    if not files:
-        logger.debug("No files provided, returning None")
-        return None
-
-    # Use configurable filesystem provider for F-016 requirements
-    fs_provider = filesystem_provider or StandardFilesystemProvider()
-    logger.debug(f"Using filesystem provider: {type(fs_provider).__name__}")
-    
-    try:
-        def get_mtime(file_path: str) -> float:
-            """Get modification time using configurable provider."""
-            try:
-                path = Path(file_path)
-                if not fs_provider.exists(path):
-                    logger.warning(f"File does not exist: {file_path}")
-                    if test_mode:
-                        return 0.0  # Return minimal time for non-existent files in test mode
-                    raise FileNotFoundError(f"File not found: {file_path}")
-                
-                stat_result = fs_provider.stat(path)
-                mtime = stat_result.st_mtime
-                logger.debug(f"File {file_path} has mtime: {mtime}")
-                return mtime
-                
-            except Exception as e:
-                logger.error(f"Error getting mtime for {file_path}: {e}")
-                if test_mode:
-                    return 0.0  # Return minimal time for error cases in test mode
-                raise
-        
-        latest_file = max(files, key=get_mtime)
-        logger.debug(f"Latest file determined: {latest_file}")
-        return str(latest_file)
-        
-    except Exception as e:
-        logger.error(f"Error in get_latest_file: {e}")
-        if test_mode:
-            logger.warning("Test mode enabled, returning first file despite error")
-            return files[0] if files else None
-        raise
+"""Compatibility layer preserving legacy import paths for discovery APIs."""
+
+from __future__ import annotations
+
+from .discoverer import FileDiscoverer, discover_files, get_latest_file
+from .experiment import discover_experiment_manifest
+from .models import FileInfo, FileManifest, FileStatistics
+
+__all__ = [
+    "FileDiscoverer",
+    "discover_files",
+    "get_latest_file",
+    "discover_experiment_manifest",
+    "FileInfo",
+    "FileManifest",
+    "FileStatistics",
+]
